convertPheno_server/pyproject.toml
[tool.poetry]
name = "convertpheno-server"
version = "0.1.0"
description = ""
authors = ["Ivo Leist"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.10.6"
anytree = "^2.8.0"
cryptography = "37.0.4"
flask-restx = "1.0.5"
flask = "2.2.3"
flask-cors = "^3.0.10"
flask-sqlalchemy = "^3.0.3"
genson = "^1.2.2"
jsonschema = "^4.17.3"
psycopg2-binary = "^2.9.6"
pyjwt = "^2.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-html = "^0.4.3"
requests = "^2.30.0"
black = "^23.3.0"
ruff = "^0.0.265"
pylint = "^2.17.4"

[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
